Extract CommandsOptions to its own file
